feat(homepage): add defaultCover to GAME_CATALOG entries
Moves cover defaults from a local map in cover.ts into GAME_CATALOG so
there is a single source of truth. Splits Cover type into cover-type.ts
to break the registry ↔ cover import cycle. Updates all inline
GameCatalogEntry fixtures in stories and tests to include defaultCover.
