+++ src/games/cover-type.ts
@@ -1,0 +1,12 @@
+export type Cover =
+  | {
+      kind: 'emoji';
+      emoji: string;
+      gradient?: [string, string];
+    }
+  | {
+      kind: 'image';
+      src: string;
+      alt?: string;
+      background?: string;
+    };
