+++ src/games/cover.ts
@@ -1,41 +1,15 @@
-export type Cover =
-  | {
-      kind: 'emoji';
-      emoji: string;
-      gradient?: [string, string];
-    }
-  | {
-      kind: 'image';
-      src: string;
-      alt?: string;
-      background?: string;
-    };
-
-const DEFAULT_COVERS: Record<string, Cover> = {
-  'word-spell': {
-    kind: 'emoji',
-    emoji: '🔤',
-    gradient: ['#fde68a', '#fb923c'],
-  },
-  'number-match': {
-    kind: 'emoji',
-    emoji: '🔢',
-    gradient: ['#bae6fd', '#6366f1'],
-  },
-  'sort-numbers': {
-    kind: 'emoji',
-    emoji: '📊',
-    gradient: ['#bbf7d0', '#10b981'],
-  },
-};
+import { GAME_CATALOG } from './registry';
+import type { Cover } from './cover-type';
 
 export const resolveDefaultCover = (gameId: string): Cover => {
-  const cover = DEFAULT_COVERS[gameId];
-  if (!cover) throw new Error(`unknown gameId: ${gameId}`);
-  return cover;
+  const entry = GAME_CATALOG.find((g) => g.id === gameId);
+  if (!entry) throw new Error(`unknown gameId: ${gameId}`);
+  return entry.defaultCover;
 };
 
 export const resolveCover = (
   doc: { cover?: Cover },
   gameId: string,
 ): Cover => doc.cover ?? resolveDefaultCover(gameId);
+
+export { type Cover } from './cover-type';
