+++ src/games/registry.ts
@@ -1,3 +1,5 @@
+import type { Cover } from './cover-type';
+
 export type GameLevel = 'PK' | 'K' | '1' | '2' | '3' | '4';
 export type GameSubject = 'math' | 'reading' | 'letters';
 
@@ -7,6 +9,7 @@ export type GameCatalogEntry = {
   descriptionKey: string;
   levels: GameLevel[];
   subject: GameSubject;
+  defaultCover: Cover;
 };
 
 export const GAME_CATALOG: GameCatalogEntry[] = [
@@ -16,6 +19,11 @@ export const GAME_CATALOG: GameCatalogEntry[] = [
     descriptionKey: 'word-spell-description',
     levels: ['PK', 'K', '1'],
     subject: 'reading',
+    defaultCover: {
+      kind: 'emoji',
+      emoji: '🔤',
+      gradient: ['#fde68a', '#fb923c'],
+    },
   },
   {
     id: 'number-match',
@@ -23,6 +31,11 @@ export const GAME_CATALOG: GameCatalogEntry[] = [
     descriptionKey: 'number-match-description',
     levels: ['1', '2'],
     subject: 'math',
+    defaultCover: {
+      kind: 'emoji',
+      emoji: '🔢',
+      gradient: ['#bae6fd', '#6366f1'],
+    },
   },
   {
     id: 'sort-numbers',
@@ -30,5 +43,10 @@ export const GAME_CATALOG: GameCatalogEntry[] = [
     descriptionKey: 'sort-numbers-description',
     levels: ['K', '1', '2'],
     subject: 'math',
+    defaultCover: {
+      kind: 'emoji',
+      emoji: '📊',
+      gradient: ['#bbf7d0', '#10b981'],
+    },
   },
 ];
